refactor: move directory interfaces
