+++ src/common/interfaces/IAuth.ts
