+++ src/common/interfaces/ICourse.ts
