+++ src/common/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from './IAuth';
+export * from './ICourse';
+export * from './IFeed';
