--- src/common/interfaces/structures/index.ts
@@ -1,1 +1,0 @@
-export * from './IAuth';
