+++ src/modules/feeds/feeds.controller.ts
@@ -1,8 +1,8 @@
 import { Controller, Get, Query } from '@nestjs/common';
 import { session_userprofile } from '@prisma/client';
 import { GetUser } from 'src/common/decorators/get-user.decorator';
+import { IFeed } from 'src/common/interfaces/IFeed';
 import { toJsonFeedDetails } from 'src/common/interfaces/serializer/feeds.serializer';
-import { IFeed } from 'src/common/interfaces/structures/IFeed';
 import { FeedsService } from './feeds.service';
 
 @Controller('api/users/:userId/feeds')
